// api/src/modules/employee/constants/enums/errors.enum.ts
export enum ErrorMessage {
  EMPLOYEE_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [nhân sự].',
  LANGUAGE_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [ngôn ngữ].',
  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',
  NO_CONTENT = 'Không có dữ liệu hiển thị.',

  OPERATOR_EMPLOYEE_ERROR = 'Lưu thông tin [nhân sự] thất bại.',
  EMPLOYEE_NOT_FOUND_ERROR = '[Nhân sự] không tồn tại (id: %s)',

  FILE_NOT_FOUND_ERROR = 'Không tìm thấy file.',

  TIME_NAN_ERROR = '[Thời gian] không hợp lệ.',
}